compile: Move and rename `SimpleIntrinsics` -> `BuiltinsContext`
This will allow us to use the class for more than just math builtins
